Add Date extension
